requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
PyJWT==2.8.0
python-dotenv==1.0.1
python-nmap==0.7.1
requests==2.32.3
scapy==2.5.0
twilio==9.2.1
urllib3==2.2.2
yarl==1.9.4
